Use notebook cell factory for cell manipulation
