+++ src/client/datascience/jupyter/jupyterNotebook.ts
@@ -23,7 +23,6 @@ import { PythonInterpreter } from '../../interpreter/contracts';
 import { captureTelemetry, sendTelemetryEvent } from '../../telemetry';
 import { generateCells } from '../cellFactory';
 import { CellMatcher } from '../cellMatcher';
-import { concatMultilineStringInput, concatMultilineStringOutput, formatStreamText } from '../common';
 import { CodeSnippits, Identifiers, Telemetry } from '../constants';
 import {
     CellState,
@@ -42,6 +41,7 @@ import { LiveKernelModel } from './kernels/types';
 
 // tslint:disable-next-line: no-require-imports
 import cloneDeep = require('lodash/cloneDeep');
+import { formatStreamText, concatMultilineStringOutput, concatMultilineStringInput } from '../../../datascience-ui/common';
 
 class CellSubscriber {
     private deferred: Deferred<CellState> = createDeferred<CellState>();
