+++ src/test/datascience/common.unit.test.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 'use strict';
 import { assert } from 'chai';
-import { formatStreamText } from '../../client/datascience/common';
+import { formatStreamText } from '../../datascience-ui/common';
 
 suite('Data Science Common Tests', () => {
     test('formatting stream text', async () => {
